requirements.txt
openai
langchain
pinecone-client
tiktoken
python-dotenv
requests
bs4
streamlit